+++ src/transport/subscription-worker/subscription-worker-middleware.ts
@@ -9,10 +9,12 @@
 
 import { CancellationController, TransportRequest } from '../../core/types/transport-request';
 import { TransportResponse } from '../../core/types/transport-response';
+import { TokenManager } from '../../core/components/token_manager';
 import * as PubNubSubscriptionWorker from './subscription-worker';
 import { PubNubAPIError } from '../../errors/pubnub-api-error';
 import StatusCategory from '../../core/constants/categories';
 import { Transport } from '../../core/interfaces/transport';
+import * as PAM from '../../core/types/api/access-manager';
 
 // --------------------------------------------------------
 // ------------------------ Types -------------------------
@@ -73,6 +75,11 @@ type PubNubMiddlewareConfiguration = {
   heartbeatInterval?: number;
 
   /**
+   * REST API endpoints access tokens manager.
+   */
+  tokenManager?: TokenManager;
+
+  /**
    * Platform-specific transport for requests processing.
    */
   transport: Transport;
@@ -107,6 +114,17 @@ export class SubscriptionWorkerMiddleware implements Transport {
    * Whether subscription worker has been initialized and ready to handle events.
    */
   subscriptionWorkerReady: boolean = false;
+
+  /**
+   * Map of base64-encoded access tokens to their parsed representations.
+   */
+  accessTokensMap: Record<
+    string,
+    {
+      token: string;
+      expiration: number;
+    }
+  > = {};
 
   constructor(private readonly configuration: PubNubMiddlewareConfiguration) {
     this.workerEventsQueue = [];
@@ -153,7 +171,9 @@ export class SubscriptionWorkerMiddleware implements Transport {
         this.callbacks!.set(req.identifier, { resolve, reject });
 
         // Trigger request processing by Service Worker.
-        this.scheduleEventPost(sendRequestEvent);
+        this.parsedAccessTokenForRequest(req)
+          .then((accessToken) => (sendRequestEvent.token = accessToken))
+          .then(() => this.scheduleEventPost(sendRequestEvent));
       }),
       controller,
     ];
@@ -325,6 +345,74 @@ export class SubscriptionWorkerMiddleware implements Transport {
         reject(new PubNubAPIError(message, category, 0, new Error(message)));
       }
     }
+  }
+
+  /**
+   * Get parsed access token object.
+   *
+   * @param req - Transport request which may contain access token for processing.
+   *
+   * @returns Object with stringified access token information and expiration date information.
+   */
+  private async parsedAccessTokenForRequest(req: TransportRequest) {
+    const accessToken = req.queryParameters ? ((req.queryParameters.auth as string) ?? '') : undefined;
+    if (!accessToken) return undefined;
+    else if (this.accessTokensMap[accessToken]) return this.accessTokensMap[accessToken];
+
+    return this.stringifyAccessToken(accessToken).then(([token, stringifiedToken]) => {
+      if (!token || !stringifiedToken) return undefined;
+
+      return (this.accessTokensMap = {
+        [accessToken]: { token: stringifiedToken, expiration: token.timestamp * token.ttl * 60 },
+      })[accessToken];
+    });
+  }
+
+  /**
+   * Stringify access token content.
+   *
+   * Stringify information about resources with permissions.
+   *
+   * @param tokenString - Base64-encoded access token which should be parsed and stringified.
+   *
+   * @returns Tuple with parsed access token and its stringified content hash string.
+   */
+  private async stringifyAccessToken(tokenString: string): Promise<[PAM.Token | undefined, string | undefined]> {
+    if (!this.configuration.tokenManager) return [undefined, undefined];
+    const token = this.configuration.tokenManager.parseToken(tokenString);
+    if (!token) return [undefined, undefined];
+
+    // Translate permission to short string built from first chars of enabled permission.
+    const stringifyPermissions = (permission: PAM.Permissions) =>
+      Object.entries(permission)
+        .filter(([_, v]) => v)
+        .map(([k]) => k[0])
+        .sort()
+        .join('');
+
+    const stringifyResources = (resource: PAM.Token['resources']) =>
+      resource
+        ? Object.entries(resource)
+            .sort(([a], [b]) => a.localeCompare(b))
+            .map(([type, entries]) =>
+              Object.entries(entries || {})
+                .sort(([a], [b]) => a.localeCompare(b))
+                .map(([name, perms]) => `${type}:${name}=${perms ? stringifyPermissions(perms) : ''}`)
+                .join(','),
+            )
+            .join(';')
+        : '';
+
+    let accessToken = [stringifyResources(token.resources), stringifyResources(token.patterns), token.authorized_uuid]
+      .filter(Boolean)
+      .join('|');
+
+    if (typeof crypto !== 'undefined' && crypto.subtle) {
+      const hash = await crypto.subtle.digest('SHA-256', new TextEncoder().encode(accessToken));
+      accessToken = String.fromCharCode(...new Uint8Array(hash));
+    }
+
+    return [token, typeof btoa !== 'undefined' ? btoa(accessToken) : accessToken];
   }
 
   /**